feat: adding landing page and css imporvement
